cleanup old diff
